I reorganized the routes into separate files depending on their uses. I move the authentication routes into routes/auth, general routes into routes/home, and the monopoly route into routes/monopoly.
I put monopoly_sockets.py (which is all of the sockets) into routes/monopoly, since only monopoly uses socketio.
