--- templates/creation.html
@@ -1,44 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Creation{% endblock %}
-
-
-{% block content %}
-
-
-
-
-
-
-<div class="box">
-    <div>
-        <h1>Elements</h1>
-        <ul>
-            {% for player in players %}
-            <li>
-                {{player}}
-                <button onclick="deletePerson({{player.id}})">Delete</button>
-                <a href="{{url_for('details', id=player.id)}}">Details</a>
-            </li>
-            {% endfor %} {% for property in properties %}
-
-            <li>{{property.title}}</li>
-
-            {% endfor %}
-        </ul>
-    </div>
-
-    <div>
-        <h3>Buy</h3>
-        <form action="{{url_for('creation')}}" method="POST">
-            <input type="text" name="buyer" placeholder="Player" required="required" /><br />
-            <input type="text" name="sold" placeholder="Property" required="required" /><br />
-            <input type="submit" name="buy" value="Create" /><br />
-        </form>
-
-        <link rel="stylesheet" href="{{url_for('static', filename='css/creation.css')}}" />
-
-        <script type="text/javascript" src="{{url_for('static', filename='js/creation.js')}}"></script>
-
-        {% endblock %}
-    </div>
-</div>
